check for strawberry or marshmallows in menu
For now only print message.  Shortly we'll choose which decoration gets
used in following clicks on the cake.

This is not perfect but it's fairly close and nothing too bad happens if
the player clicks just next to the strawberry and the game thinks it is
actually on the strawberry.  The numbers in the tests were found by
clicking on the screen and noting the numbers printed by the

    'button 1 clicked at %d, %d' % (click_x, click_y)

message.